제발좀 commit좀 하자

+++ spartamarket/products/views.py
@@ -1,6 +1,9 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from django.http import HttpResponse
 from .models import Article
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_http_methods
+from .forms import ArticleForm
 
 
 def articles(request):
@@ -9,47 +12,59 @@ def articles(request):
     return render(request, "articles.html", context)
 
 
-def new(request):
-    return render(request, "new.html")
-
-
+@login_required
 def create(request):
+    if request.method == "POST":
+        form = ArticleForm(request.POST)
+        if form.is_valid():
+            article = form.save()
+            return redirect("products:detail", article.id)
+    else:
+        form = ArticleForm()
 
-    title = request.POST.get("title")
-    content = request.POST.get("content")
-
-    article = Article(title=title, content=content)
-    article.save()
-    context = {
-        "article": article,
-    }
-    return redirect("products:articles")
+    context = {"form": form}
+    return render(request, "create.html", context)
 
 
 def detail(request, pk):
-    article = Article.objects.get(id=pk)
+    article = get_object_or_404(Article, pk=pk)
     context = {"article": article}
     return render(request, "detail.html", context)
 
 
+@require_http_methods(["POST"])
 def delete(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        article.delete()
     return redirect("products:articles")
 
 
-def edit(request, pk):
-    article = Article.objects.get(pk=pk)
-    context = {"article": article}
-    return render(request, "edit.html", context)
-
-
 def update(request, pk):
-    title = request.POST.get("title")
-    content = request.POST.get("content")
+    article = get_object_or_404(Article, pk=pk)
+    if request.method == "POST":
+        form = ArticleForm(request.POST, instance=article)
+        if form.is_valid():
+            article = form.save()
+            return redirect("products:detail", article.pk)
+    else:
+        form = ArticleForm(instance=article)
+    context = {
+        "form": form,
+        "article": article,
+    }
+    return render(request, "update.html", context)
 
-    article = Article.objects.get(pk=pk)
-    article.content = content
-    article.title = title
-    article.save()
-    return redirect("products:detail", article.pk)
+
+@require_http_methods(["POST"])
+def like(request, pk):
+    if request.user.is_authenticated:
+        article = get_object_or_404(Article, pk=pk)
+        if article.like_users.filter(pk=request.user.pk).exists():
+            article.like_users.remove(request.user)
+        else:
+            article.like_users.add(request.user)
+    else:
+        return redirect("accounts:login")
+
+    return redirect("articles:articles")
